+++ VulkanEngine/Main.cpp
@@ -88,9 +88,9 @@ namespace game {
 	Engine												engine;
 	Camera												camera;
 	bool												firstMouse							= true;
-	float												lastX								= WIDTH / 2;
-	float												lastY								= HEIGHT / 2;
-	float												DELTATIME;
+	double												lastX								= WIDTH / 2;
+	double												lastY								= HEIGHT / 2;
+	double												DELTATIME;
 	GLFWwindow*											pWindow;
 	const unsigned int									MAX_FRAMES_IN_FLIGHT				= 2;
